Fix for test.py

+++ user_service/views.py
@@ -1,12 +1,12 @@
-from .models import Client
+from user_service.models import Client
 from django.http import HttpResponse, JsonResponse
-from .serializers import ClientSerializer
 from rest_framework import permissions, generics
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework.generics import CreateAPIView
 from rest_framework.authtoken.models import Token
 from django.shortcuts import render
+from .serializers import ClientSerializer
 from django.shortcuts import get_object_or_404
 from rest_framework.decorators import api_view
 from rest_framework.status import (
@@ -15,7 +15,6 @@ from rest_framework.status import (
     HTTP_404_NOT_FOUND,
     HTTP_400_BAD_REQUEST,
 )
-
 class ClientList(generics.ListCreateAPIView):
     queryset = Client.objects.all()
     serializer_class = ClientSerializer
@@ -30,7 +29,7 @@ def register_client(request):
     data = {}
 
     if serializer.is_valid():
-        client = serializer.save()
+        order = serializer.save()
         data['response'] = 'Usuário registrado com sucesso'
 
     else:
@@ -38,19 +37,28 @@ def register_client(request):
 
     return Response(data)
 
-@api_view(['DELETE'])
-def delete_client(request, registro):
-    try:
-        client = Client.objects.get(pk=registro)
-    except Client.DoesNotExist:
-        return HttpResponse(status=404)
-    client.delete()
-    return HttpResponse(status=204)
-
 @api_view(["POST"])
 def list_clients(request):
     clients = Client.objects.all().values()
     return Response(data=clients)
+
+@api_view(['DELETE'])
+def delete_client(request, registro):
+
+    # If request is valid
+    client = Client.objects.get(pk=registro)
+    if (registro == None):
+        return Response({'error': 'Formulário inválido.'},
+                                status=HTTP_400_BAD_REQUEST)
+    # If order exist
+    try:
+        client = Client.objects.get(pk=registro)
+    except:
+        return Response({'error': 'Produto não existe.'},
+                                status=HTTP_404_NOT_FOUND)
+    client.delete()
+    return HttpResponse(status=204)
+
 
 @api_view(["POST"])
 def edit_client(request, registro):
